A quick fix of QCLContextGL::create() method - implementation for Windows OpenGL context sharing was missing. The code has yet to be tested on other platforms (at least Mac and X11) and with more than one platform and more than one OpenCL device.
In the original implementation is a strange problem during kernel compilation.
It manifests itself by a opening a terminal window for each clBuildProgram call.

I observed that the problem was caused by having a CPU device in between devices
of QCLContextGL context, which later during compilation for all devices present in
the context manifested itself in the described way.
When only the OpenGL device was part of the context this problem was not observed.

This has yet to be solved - it is most likely a driver bug or an API misuse,
so I will either have to find a workaround or learn the proper way to implment this.